fix: Handle tables with missing <thead> tag before attaching a DataTable
The previous code would ALWAYS insert a <thead> tag. However, some tables already have it and the final was duplicated headers. This fix checks wether a <thead> is present or not, and in case it's missing it makes a more robust search for a possible table header (which is not mandatory). If one is found, then is wrapped around a <thead> and inserted in the original table

Closes #78